canvas2d: introduce current canvas 2d emulator
NanoVG does not match the canvas2d spec fully but this is
the current level of emulation we have achieved.

This seems to work on most mobile browsers well.
Excluding Chrome which has a scaling issue because the
context is set to a strange size and gl can not be resized.